+++ hal/src/sercom/i2c/flags.rs
@@ -87,6 +87,10 @@ impl Status {
             Ok(())
         }
     }
+
+    pub fn is_idle(self) -> bool {
+        self.busstate() == BusState::Idle
+    }
 }
 
 /// Errors available for I2C transactions
@@ -98,4 +102,6 @@ pub enum Error {
     LengthError,
     Nack,
     Timeout,
+    #[cfg(feature = "dma")]
+    Dma(crate::dmac::Error),
 }