+++ hal/src/sercom/mod.rs
@@ -34,6 +34,7 @@ use core::ops::Deref;
 
 use crate::pac;
 use pac::sercom0;
+use pac::Peripherals;
 
 #[hal_cfg("sercom0-d5x")]
 use pac::Mclk as ApbClkCtrl;
@@ -50,11 +51,19 @@ pub use pad::*;
 
 pub mod i2c;
 pub mod spi;
+
+#[deprecated(
+    since = "0.18.0",
+    note = "spi_future is deprecated and will be removed in a later version of atsamd_hal. Consider using the `async` APIs available in the `spi` module as a replacement."
+)]
 pub mod spi_future;
 pub mod uart;
 
 #[cfg(feature = "dma")]
 pub mod dma;
+
+#[cfg(all(feature = "dma", feature = "async"))]
+mod async_dma;
 
 //==============================================================================
 //  Sercom
@@ -64,14 +73,38 @@ pub mod dma;
 pub trait Sercom: Sealed + Deref<Target = sercom0::RegisterBlock> {
     /// SERCOM number
     const NUM: usize;
+
     /// RX Trigger source for DMA transactions
     #[cfg(feature = "dma")]
     const DMA_RX_TRIGGER: TriggerSource;
+
     /// TX trigger source for DMA transactions
     #[cfg(feature = "dma")]
     const DMA_TX_TRIGGER: TriggerSource;
+
+    #[cfg(feature = "async")]
+    type Interrupt: crate::async_hal::interrupts::InterruptSource;
+
     /// Enable the corresponding APB clock
     fn enable_apb_clock(&mut self, ctrl: &ApbClkCtrl);
+
+    /// Get a reference to the sercom from a
+    /// [`Peripherals`] block
+    fn reg_block(peripherals: &mut Peripherals) -> &crate::pac::sercom0::RegisterBlock;
+
+    /// Get a reference to this [`Sercom`]'s associated RX Waker
+    #[cfg(feature = "async")]
+    #[inline]
+    fn rx_waker() -> &'static embassy_sync::waitqueue::AtomicWaker {
+        &crate::sercom::async_api::RX_WAKERS[Self::NUM]
+    }
+
+    /// Get a reference to this [`Sercom`]'s associated TX Waker
+    #[cfg(feature = "async")]
+    #[inline]
+    fn tx_waker() -> &'static embassy_sync::waitqueue::AtomicWaker {
+        &crate::sercom::async_api::TX_WAKERS[Self::NUM]
+    }
 }
 
 macro_rules! sercom {
@@ -83,15 +116,33 @@ macro_rules! sercom {
             impl Sealed for [< Sercom $N >] {}
             impl Sercom for [< Sercom $N >] {
                 const NUM: usize = $N;
+
                 #[cfg(feature = "dma")]
                 const DMA_RX_TRIGGER: TriggerSource = TriggerSource::[< Sercom $N Rx >];
+
                 #[cfg(feature = "dma")]
                 const DMA_TX_TRIGGER: TriggerSource = TriggerSource::[< Sercom $N Tx >];
+
+                #[cfg(feature = "async")]
+                #[hal_cfg(any("sercom0-d11", "sercom0-d21"))]
+                type Interrupt = $crate::async_hal::interrupts::[< SERCOM $N >];
+
+                #[cfg(feature = "async")]
+                #[hal_cfg("sercom0-d5x")]
+                type Interrupt = $crate::async_hal::interrupts::[< SERCOM $N >];
+
                 #[inline]
                 fn enable_apb_clock(&mut self, ctrl: &ApbClkCtrl) {
                     ctrl.$apbmask().modify(|_, w| w.[< sercom $N _>]().set_bit());
                 }
+
+                #[inline]
+                fn reg_block(peripherals: &mut Peripherals) -> &crate::pac::sercom0::RegisterBlock {
+                    &*peripherals.[< sercom $N >]
+                }
             }
+
+
         }
     };
 }
@@ -139,3 +190,31 @@ sercom!(apbdmask, 6);
 
 #[hal_cfg("sercom7-d5x")]
 sercom!(apbdmask, 7);
+
+// Reserve space for the max number of SERCOM peripherals based on chip type,
+// even though some wakers may not be used on some chips if they actually don't
+// exist on variant's hardware
+#[hal_cfg("sercom0-d11")]
+#[cfg(feature = "async")]
+const NUM_SERCOM: usize = 3;
+
+#[hal_cfg("sercom0-d21")]
+#[cfg(feature = "async")]
+const NUM_SERCOM: usize = 6;
+
+#[hal_cfg("sercom0-d5x")]
+#[cfg(feature = "async")]
+const NUM_SERCOM: usize = 8;
+
+#[cfg(feature = "async")]
+pub(super) mod async_api {
+    use embassy_sync::waitqueue::AtomicWaker;
+
+    #[allow(clippy::declare_interior_mutable_const)]
+    const NEW_WAKER: AtomicWaker = AtomicWaker::new();
+    /// Waker for a RX event. By convention, if a SERCOM has only one type of
+    /// event (ie, I2C), this the waker to be used.
+    pub(super) static RX_WAKERS: [AtomicWaker; super::NUM_SERCOM] = [NEW_WAKER; super::NUM_SERCOM];
+    /// Waker for a TX event.
+    pub(super) static TX_WAKERS: [AtomicWaker; super::NUM_SERCOM] = [NEW_WAKER; super::NUM_SERCOM];
+}